sitemap_server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.3.4
    # via starlette
asgiref==3.4.1
    # via uvicorn
certifi==2021.10.8
    # via sentry-sdk
click==8.0.3
    # via uvicorn
fastapi==0.70.0
    # via -r requirements.in
fluent==0.10.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==3.3
    # via anyio
lxml==4.6.3
    # via -r requirements.in
mysqlclient==2.0.3
    # via -r requirements.in
pyaml==21.10.1
    # via -r requirements.in
pydantic[dotenv]==1.8.2
    # via
    #   -r requirements.in
    #   fastapi
python-dotenv==0.19.1
    # via pydantic
pyyaml==6.0
    # via pyaml
sentry-sdk==1.4.3
    # via -r requirements.in
sniffio==1.2.0
    # via anyio
starlette==0.16.0
    # via fastapi
typing-extensions==3.10.0.2
    # via pydantic
urllib3==1.26.7
    # via sentry-sdk
uvicorn==0.15.0
    # via -r requirements.in
